requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
django-referrer-policy==1.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
frozenlist==1.4.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.7
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain==0.3.3
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.135
multidict==6.1.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.7
packaging==24.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.32.0
whitenoise==6.8.2
yarl==1.15.3
youtube-transcript-api==0.6.2
